+++ src/core/object/select/CHANGELOG.md
@@ -9,6 +9,16 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.48.0 (2021-05-18)
+
+#### :rocket: New Feature
+
+* Added support for Iterable objects
+
+#### :bug: Bug Fix
+
+* Bugfixes
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature